Go refactor: merge similar functions
